# NF: `clone` a faster `install` for most cases