--- a/datalad/interface/__init__.py
+++ b/datalad/interface/__init__.py
@@ -23,6 +23,7 @@ _group_dataset = (
         # source module, source object[, dest. cmdline name[, dest python name]]
         # src module can be relative, but has to be relative to the main 'datalad' package
         ('datalad.distribution.create', 'Create'),
+        ('datalad.distribution.clone', 'Clone'),
         ('datalad.distribution.install', 'Install'),
         ('datalad.distribution.get', 'Get'),
         ('datalad.distribution.add', 'Add'),