… datalad/distribution/get.py
… datalad/distribution/tests/test_clone.py
… datalad/distribution/tests/test_get.py
… datalad/distribution/tests/test_install.py
… datalad/interface/base.py
… datalad/interface/common_opts.py
… datalad/interface/tests/test_utils.py
… datalad/metadata/search.py
… datalad/support/gitrepo.py
… datalad/tests/utils_testdatasets.py
… docs/source/cmdline.rst