Check `datalad/distribution/tests/test_clone.py` for converted install tests that represent the scope of `clone`. Major change is dramatically reduced flexibility wrt input arguments (see #1396). This leads to about ~10% of speed up for our bread-and-butter call, i.e. install one dataset (recursive or not) from one source.

It looks like we could switch to `clone` for pretty much all use in the tests, except for the test of `install`. Even if `install` represents only a fraction of the actual runtime, we are calling it a lot, so it should make things a bit faster.

Here is a demo:

````
In [1]: %run clone_speed_demo.py

# disregard the first run -- testrepos are built here (see below for the code)
In [2]: %timeit test_clone_speed()
1 loop, best of 3: 3.63 s per loop

In [3]: %timeit test_install_speed()
1 loop, best of 3: 4.02 s per loop

In [4]: %timeit test_install_speed()
1 loop, best of 3: 4.04 s per loop

In [5]: %timeit test_clone_speed()
1 loop, best of 3: 3.63 s per loop

In [6]: %timeit test_clone_speed()
1 loop, best of 3: 3.68 s per loop

In [8]: %timeit test_clone_speed()
1 loop, best of 3: 3.64 s per loop
````

Here is the code. It performs a recursive install to get a lower bound estimate. Both commands use `get` for the recursion part, albeit in a slightly different way. Not going recursive yields a ~12% speed increase.

````
from datalad.tests.utils import with_testrepos
from datalad.tests.utils import with_tempfile
from datalad.api import clone
from datalad.api import install
from datalad.api import Dataset


@with_testrepos('.*nested_submodule.*', flavors=['local-url', 'network', 'local'])
@with_tempfile(mkdir=True)
def test_clone_speed(src, path):
    Dataset(path)

    clone(src, path, recursive=True,
          result_xfm='datasets', return_type='item-or-list')


@with_testrepos('.*nested_submodule.*', flavors=['local-url', 'network', 'local'])
@with_tempfile(mkdir=True)
def test_install_speed(src, path):
    Dataset(path)

    install(path, source=src, recursive=True)
````