@yarikoptic 2017-03-20
Not sure if we sold be solving performance issues by expanding API... I am already confused by having both install and clone