@bpoldrack 2017-03-20
Since you are familiar with git, `clone` should not exactly be a confusing thing. So, if you want to clone, use `clone. If you want a lot of guessing magic and simultaneously add things to superdatasets and so on, use `install`.