@mih 2017-03-20
One more datapoint: For `nosetests3 -s datalad/tests/test_testrepos.py:test_make_studyforrest_mockup` alone I get an ~8% speedup (best of three iterations) by doing `s/install/clone/g`. In absolute numbers this is ~10s.