@yarikoptic 2017-03-20
```
Since you are familiar with git, clone should not exactly be a confusing thing. So, if you want to clone, use clone. If you want a lot of guessing magic and simultaneously add things to superdatasets and so on, use install.
```
that is my point -- how would I know if I need any magic or not when I am cloning/installing a remote source?  also moving testing of install in various scenarios as it is done now into testing through a new command could potentially leave install less tested and thus magic break more often.