@yarikoptic 2017-03-20
I feel deficient of having no "my clone" to move stuff to ;-)  seriously -- what would be the "function in line" of the new API command?  is it for a pro-user (or tests) to avoid using `install` when he could just use `clone` because it is faster?  or... ?