@mih 2017-03-21
@bpoldrack test failure for a direct mode test is the last issue. could you please have a quick look. there still seems to be a necessity that i am not aware off. thanks.